completed all datasets

+++ dataset/polblog_data.py
@@ -27,7 +27,7 @@ class PolBlogDataFrame(object):
 
         # Create the membership variables
         self.X = torch.cat([torch.from_numpy(dfn[col].values.reshape(-1, 1)) for col in feature_cols], dim=1, )
-        self.edge_index = torch.cat([torch.from_numpy(dfe[col].values.reshape(-1, 1).astype(np.int32)) for col in ["source", "target"]],dim=1).T
+        self.edge_index = torch.cat([torch.from_numpy(dfe[col].values.reshape(-1, 1).astype(np.int32)) for col in ["source", "target"]],dim=1).T.long()
         self.group_col = feature_cols.index(group_col)
 
     def get_grouped_col(self):
